End of Section 18

+++ client/src/app/messages/messages.component.ts
@@ -1,3 +1,4 @@
+import { ConfirmService } from './../_services/confirm.service';
 import { Pagination } from './../_models/pagination';
 import { Component, OnInit } from '@angular/core';
 import { Message } from '../_models/message';
@@ -16,7 +17,7 @@ pageNumber=1;
 pageSize=5;
 loading= false;
 
-  constructor(private messageService: MessageService) { }
+  constructor(private messageService: MessageService, private confirmService: ConfirmService) { }
 
   ngOnInit(): void {
     this.loadMessages();
@@ -32,9 +33,18 @@ loading= false;
   }
   deleteMessage(id:number)
   {
-    this.messageService.deleteMessage(id).subscribe(() => {
-      this.messages.splice(this.messages.findIndex(m => m.id=id),1);
-    });
+
+    this.confirmService.confirm('Confirm delete message', 'This cannot be undone').subscribe(result =>
+      {
+        if(result){
+
+          this.messageService.deleteMessage(id).subscribe(() => {
+            this.messages.splice(this.messages.findIndex(m => m.id=id),1);
+          });
+
+        }
+      });
+
 
   }
